chore: consolidate scripts to single source of truth and purge binaries
